refactor get_cooccurrence_matrix and get_annotation_pairs_for_multiple_annotators to _metrics.py
