+++ gitma/project.py
@@ -2,8 +2,6 @@ import subprocess
 import os
 import json
 import textwrap
-from collections import defaultdict, Counter
-from itertools import combinations
 from typing import Any, Dict, List, Tuple, Union, Generator
 
 import gitlab
@@ -21,7 +19,9 @@ from gitma.tag import Tag
 from gitma._write_annotation import write_annotation_json
 from gitma._gold_annotation import create_gold_annotations
 from gitma._vizualize import plot_interactive, plot_annotation_progression
-from gitma._metrics import get_annotation_pairs, get_iaa_data, get_confusion_matrix, gamma_agreement, EmptyAnnotation
+from gitma._metrics import get_annotation_pairs, get_iaa_data, get_confusion_matrix, gamma_agreement, EmptyAnnotation, \
+    get_annotation_pairs_for_multiple_annotators, get_cooccurence_matrix
+
 
 def load_gitlab_project(
         gitlab_access_token: str,
@@ -973,122 +973,6 @@ class CatmaProject:
 
         return self._return_iaa_result(annotation_task.kappa, "Cohen's Kappa", level, confusion_matrix, verbose)
 
-    def get_cooccurence_matrix(self, annotationdata: set) -> pd.DataFrame:
-        """Generates cooccurence matrix for annotation data
-
-        Args:
-            annotationdata (Set[List[coderid,itemid,tag]]): List of overlapping annotations as a set.
-
-        Returns:
-            pd.DataFrame: Cooccurence matrix as pandas data frame.
-        """
-
-        # group tags by item
-        items_tags_grouped = defaultdict(list)
-        for coderid, itemid, tag in annotationdata:
-            items_tags_grouped[itemid].append(tag)
-
-        # count co-coccurences of tags
-        cooccurence_counter = Counter()
-        for tags in items_tags_grouped.values():
-            for pair in combinations(sorted(tags), 2):  # all possible combinations of two elements
-                cooccurence_counter[pair] += 1
-
-        # matrix
-        all_tags = sorted(set(tag for _, _, tag in annotationdata))
-
-        matrix = pd.DataFrame(0, index=all_tags, columns=all_tags)
-
-        for (a, b), count in cooccurence_counter.items():
-            matrix.loc[a, b] = count
-            matrix.loc[b, a] = count
-
-        return matrix
-
-    def get_annotation_pairs_for_multiple_annotators(
-            self,
-            ac_names: list = [],
-            tag_filter: list = [],
-            filter_both_ac: bool = True,
-            include_empty_annotations: bool = True,
-            property_filter: str = None,
-            verbose: bool = True,
-    ) -> set:
-        """
-        Get annotation data in the NLTK format for IAA calculation for two or more annotators without duplicate pairs.
-        Args:
-            ac_names (list): List of annotation collection names to include in the IAA calculation. If empty, all ACs in the project will be used.
-            tag_filter (list): List of tags that should be included for iaa calculation. If empty, all tags will be used.
-            filter_both_ac (bool): Whether to apply tag_filter on both ACs in the pair or just on the first AC. Default is True.
-            include_empty_annotations (bool): Whether to include empty annotations in the IAA data. If `False`, only annotations with a matching annotation\
-             in the second collection are included. Default is True.
-            property_filter (str, optional): Property to filter by as a string with the property name. If None, all properties will be used.\
-            verbose (bool, optional): Whether to print results to stdout. Defaults to `True`.
-        Returns:
-            Set of annotation tuples in [NLTK format](https://www.nltk.org/api/nltk.metrics.agreement.html#nltk.metrics.agreement.AnnotationTask.__init__) for IAA calculation.
-        """
-
-        if verbose:
-            if ac_names == []:
-                print("No annotation collection names provided, using all ACs in the project.")
-                ac_names = list(self.ac_dict.keys())
-            else:
-                print("Using provided annotation collection names: ", ac_names)
-
-            if tag_filter == []:
-                print("No tag filter provided, using all annotations.")
-            else:
-                print(f"Using provided tag filter: {tag_filter}, with filter_both_ac set to {filter_both_ac}.")
-
-            if not property_filter:
-                print("No property filter provided, using all annotations.")
-            else:
-                print(f"Using provided property filter: {property_filter}.")
-
-        # Get annotation collection combinations and enumerate them for IAA calculation
-        # enumerate for mapping back to original indices after getting annotation pairs
-        ac_names_enum = list(enumerate(ac_names))
-
-        # All pairwise combinations of ACs
-        ac_combinations = list(combinations(range(len(ac_names)), 2))
-
-        # Set to store all matching annotations. As we do pairwise comparison, we do not want to keep duplicates from the pairwise comparison
-        pairwise_iaa_data = set()
-
-        for ac_pair in ac_combinations:
-            ac_first_index = ac_pair[0]  # this is the enumerated index of the first AC in the pair, in current setup, it is always 0th index since we are comparing the first AC against the rest of ACs.
-            ac_second_index = ac_pair[1]  # this is the enumerated index of the second AC in pairwise comparison, could be 1,2,3,...nth
-            ac_first_name = ac_names_enum[ac_first_index][1]  # name of the first AC in the pair
-            ac_second_name = ac_names_enum[ac_second_index][1]  # name of the second AC in the pair
-
-            # Get annotation pairs for the current AC combination
-            annotation_pairs = get_annotation_pairs(
-                self.ac_dict[ac_first_name],
-                self.ac_dict[ac_second_name],
-                filter_both_ac=filter_both_ac,
-                tag_filter=tag_filter,
-                property_filter=property_filter,
-                verbose=verbose
-            )
-
-            # Converts gitma annotation pairs to IAA data format (Coder, Item, Label)
-            iaa_data = list(
-                get_iaa_data(annotation_pairs,
-                             include_empty_annotations=include_empty_annotations)
-            )
-
-            # Map the coder indices in IAA data back to the original AC indices. Only first value in the tuple is changed, which is the coder index
-            coder_map = {
-                0: ac_first_index,
-                1: ac_second_index
-            }
-
-            coderid_mapped_iaa_data = [(coder_map[coder], item, label) for coder, item, label in iaa_data]
-
-            # add the IAA data for the current AC combination to final set
-            pairwise_iaa_data.update(coderid_mapped_iaa_data)
-
-        return pairwise_iaa_data
 
     # TODO: implement level parameter
     def calculate_krippendorffs_alpha(
@@ -1126,17 +1010,18 @@ class CatmaProject:
             distance_function = binary_distance
 
         # get matching annotation pairs
-        annotation_pairs = self.get_annotation_pairs_for_multiple_annotators(
+        annotation_pairs = get_annotation_pairs_for_multiple_annotators(
+            ac_dict=self.ac_dict,
             ac_names=ac_names,
             tag_filter=tag_filter,
             filter_both_ac=filter_both_ac,
             include_empty_annotations=include_empty_annotations,
             property_filter=property_filter,
-            verbose=False
+            verbose=verbose
         )
 
         annotation_task = AnnotationTask(data=annotation_pairs, distance=distance_function)
-        co_matrix = self.get_cooccurence_matrix(annotation_pairs)
+        co_matrix = get_cooccurence_matrix(annotation_pairs)
 
         return self._return_iaa_result(annotation_task.alpha, "Krippendorff's Alpha", 'tag', co_matrix, verbose, matrix_type="Co-occurrence matrix")
